Move buffers from worktree to project
